move animation for PageArrow to its own file
